closes #15.
the 'second' login test was duplicated, comments and all, and the async
nature of the test meant that the duplicated 'second login' block got fired
first, causing the actual first login to fail.

the weirdness of async tests with the upstream expresso caused the error
to be reported as the last test, which actually had nothing to do with
it.